requirements_server.txt
evidently>=0.4.13
fastapi>=0.109.0
google-cloud-storage>=2.14.0
google-cloud-storage>=2.14.0
lightning>=2.1.3
pandas>=1.5.3
Pillow>=9.4.0
python-multipart>=0.0.6
scikit-learn>=1.2.1
timm>=0.9.12
torch>=2.1.0
transformers>=4.36.2
uvicorn>=0.25.0
